% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_About.R
\name{about}
\alias{about}
\title{about server}
\usage{
about(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object}
}
\value{
none
}
\description{
about server
}
