% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_About.R
\name{about_UI}
\alias{about_UI}
\title{about_UI - information module}
\usage{
about_UI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
information module
}
