% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Administration.R
\name{admin_action_datatable}
\alias{admin_action_datatable}
\title{action search module - server}
\usage{
admin_action_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
access to actions}
}
\value{
none
}
\description{
action search module - server
}
