% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Administration.R
\name{admin_pcehr_datatable}
\alias{admin_pcehr_datatable}
\title{myHealth (pcehr) search module - server}
\usage{
admin_pcehr_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
access to documents}
}
\value{
none
}
\description{
myHealth (pcehr) search module - server
}
