% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Administration.R
\name{admin_result_datatable}
\alias{admin_result_datatable}
\title{result management module - server}
\usage{
admin_result_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
access to appointments lists, results, correspondence and EMR database}
}
\value{
none
}
\description{
with appointments
}
