% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Administration.R
\name{administration}
\alias{administration}
\title{administration user interface}
\usage{
administration(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object}
}
\value{
none
}
\description{
administration user interface
}
