% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Administration.R
\name{administration_UI}
\alias{administration_UI}
\title{administration_UI - admin module}
\usage{
administration_UI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
admin module
}
