% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Appointments.R
\name{appointments_datatableUI}
\alias{appointments_datatableUI}
\title{Appointments module - UI function}
\usage{
appointments_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Display appointments within selected range of dates and providers
}
