% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Billings.R
\name{billings_datatable}
\alias{billings_datatable}
\title{appointment list module - server}
\usage{
billings_datatable(input, output, session, appointments_billings, db)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{db}{access to database tables from Best Practice EMR}

\item{appointments_filtered_time}{reactive list of appointments attached to all billings}
}
\value{
none
}
\description{
list of appointments and billings
within selected range of dates and providers
}
