% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Billings.R
\name{billings_datatableUI}
\alias{billings_datatableUI}
\title{Appointments module - UI function}
\usage{
billings_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Display appointments within selected range of dates and providers
}
