% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CancerScreen.R
\name{cancerscreen_datatable}
\alias{cancerscreen_datatable}
\title{Cancer screening module server}
\usage{
cancerscreen_datatable(input, output, session, appointments_list, emr_db)
}
\arguments{
\item{input}{(as required by modules)}

\item{output}{(as required by modules)}

\item{session}{(as required by modules)}

\item{appointments_list}{reactive list of appointments to search}

\item{emr_db}{R6 object, access to Electronic Medical Record database}
}
\value{
None
}
\description{
Chronic disease management items claimed, pending or unclaimed for appointment list
}
