% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CancerScreen.R
\name{cancerscreen_datatableUI}
\alias{cancerscreen_datatableUI}
\title{Cancer screen User Interface module}
\usage{
cancerscreen_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Datatable with list of patients and cancer screening opportunities.
Includes 'printable' toggle, and selectable cancer screening dropdown.
}
