% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CDM.R
\name{cdm_datatable}
\alias{cdm_datatable}
\title{Chronic disease management list module - server}
\usage{
cdm_datatable(input, output, session, appointments_billings,
  appointments_filtered, appointments_filtered_time, appointments_list,
  diabetes_list, asthma_list, dbHistory)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{appointments_billings}{reactive. joining appointments and billings}

\item{appointments_filtered}{reactive}

\item{appointments_filtered_time}{reactive. filtered appointments list}

\item{appointments_list}{reactive. same as appointments_filtered_time, but with DOB and Age added}

\item{diabetes_list}{vector of diabetic patients with appointments}

\item{asthma_list}{vector of asthma patients with apponitments}

\item{dbhistory}{table of history items, lazy evaluation}

\item{appointments_filtered_time}{reactive list of appointments attached to all billings}

\item{db}{access to database tables from Best Practice EMR}
}
\value{
none
}
\description{
chronic disease management items claimed, pending or unclaimed for appointment list
}
