% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CDM.R
\name{cdm_datatableUI}
\alias{cdm_datatableUI}
\title{Chronic Disease Management (CDM) module - UI function}
\usage{
cdm_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Display CDM status and opportunities within selected range of dates and providers
}
