% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Conditions.R
\name{conditions_asthma_datatable}
\alias{conditions_asthma_datatable}
\title{condition asthma - server}
\usage{
conditions_asthma_datatable(input, output, session, dM)

conditions_asthma_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
access to appointments lists, visits, user interface elements
and post-natal status database}

\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
access to appointments lists, visits, user interface elements
and post-natal status database}
}
\value{
none

none
}
\description{
search for asthma patients.

search for asthma patients.
}
\details{
Attached to appointments within the defined appointment search period.

Attached to appointments within the defined appointment search period.
}
