% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Conditions.R
\name{conditions_postnatal_datatable}
\alias{conditions_postnatal_datatable}
\title{data quality module - server}
\usage{
conditions_postnatal_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
access to appointments lists, results, correspondence and EMR database}
}
\value{
none
}
\description{
with appointments
}
