% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Conditions.R
\name{conditions_UI}
\alias{conditions_UI}
\title{conditions_UI - admin module}
\usage{
conditions_UI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
conditions module
}
