% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyMeasureServer.R
\name{DailyMeasureServer}
\alias{DailyMeasureServer}
\title{Shiny app server function}
\usage{
DailyMeasureServer(input, output, session)
}
\arguments{
\item{input}{required for shiny server}

\item{output}{required for shiny server}

\item{session}{required for shiny server}
}
\value{
None
}
\description{
Shiny app server function
}
