% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyMeasureUI.R
\docType{data}
\name{DailyMeasureUI}
\alias{DailyMeasureUI}
\title{Shiny app UI object}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
DailyMeasureUI
}
\description{
Shiny app UI object
}
\keyword{datasets}
