% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\docType{data}
\name{dbConnection}
\alias{dbConnection}
\title{dbConnection class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dbConnection
}
\description{
connect using DBI or pool
}
\details{
dbConnection class
}
\section{Fields}{

\describe{
\item{\code{DBIconn}}{- connection using DBI (NULL if not connected)}

\item{\code{poolconn}}{- connection using pool (NULL is not connected)}

\item{\code{conn()}}{- connection using whichever connection is available}
}}

\section{Methods}{

\describe{
\item{\strong{connect} : connect to database}
\item{\strong{close} : close connection to database}
\item{\strong{dbSendQuery} : send query (statement) to connection}
}
}

\examples{
dbConnection$new()   # creates new object

dbConnection$connect(usepool = TRUE, RSQLite::SQLite(),
                     dbname = "mydatabase.sqlite")
                     # sets $DBIconn or $poolcon to connection
                     # both set to NULL if fail/warn

dbConnection$close() # closes all connections

query <- "UPDATE Users SET Name = ? WHERE id = ?"
data_for_sql <- list(c("John"), id)
dbConnectioin$dbSendQuery(query, data_for_sql)
                     # send parametized 'SQL query'
                     # with 'data_for_sql'

}
\keyword{datasets}
