% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_displayPreferences.R
\name{displayPreferences_UI}
\alias{displayPreferences_UI}
\title{display preferenmces module - UI}
\usage{
displayPreferences_UI(id)
}
\arguments{
\item{id}{as required by shiny modules}
}
\description{
Allow user to set or change display preferences
}
