% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\docType{data}
\name{dMeasure}
\alias{dMeasure}
\title{dMeasure class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dMeasure
}
\description{
case-finding in EMR (Best Practice) a
}
\details{
dMeasure class
}
\section{Fields}{

\describe{
\item{\code{yaml_config_filepath}}{- filepath of YAML configuration}

\item{\code{sql_config_filepath}}{- filepath of SQL configuration (NULL is not connected)}

\item{\code{local_config}}{- in-memory copy of YAML configuration}
}}

\section{Methods}{

\describe{
\item{\strong{read_configuration_filepaths} : read (or initiate) YAML/SQL DB filepaths}
\item{\strong{open_configuration_db} : open SQL configuration database}
\item{\strong{read_configuration_db} : read SQL configuration database}
}
}

\examples{

}
\keyword{datasets}
