% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration.R
\name{locations_datatable}
\alias{locations_datatable}
\title{location configuration module - server}
\usage{
locations_datatable(input, output, session, PracticeLocations, UserConfig,
  config_pool)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{PracticeLocations}{reactiveval, list of locations}

\item{UserConfig}{reactiveval, list of users who have been configured}

\item{config_pool}{reactiveval, access to configuration database}
}
\value{
count - increments with each edit of server database
}
\description{
Editable datatable with list of locations
}
