% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_location.R
\name{locations_datatableUI}
\alias{locations_datatableUI}
\title{location configuration module - UI function}
\usage{
locations_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Editable datatable with list of locations
}
