% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_logs.R
\name{logging_datatable}
\alias{logging_datatable}
\title{userconfig_resetpassword - server component
reset password of selected user}
\usage{
logging_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object, access to databases, user config}
}
\value{
nothing
}
\description{
userconfig_resetpassword - server component
reset password of selected user
}
