% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_logs.R
\name{logging_datatableUI}
\alias{logging_datatableUI}
\title{loggingUI}
\usage{
logging_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
shiny user interface element
}
\description{
enable logging, choose logging SQLite database file
}
