% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_password.R
\name{passwordConfig_server}
\alias{passwordConfig_server}
\title{password configuration module - server function}
\usage{
passwordConfig_server(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object}
}
\value{
count - increments with each GUI edit of user configuration database
}
\description{
Allow user to set or change their user password
}
