% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration.R
\name{passwordConfig_UI}
\alias{passwordConfig_UI}
\title{password configuration module - UI}
\usage{
passwordConfig_UI(id)
}
\arguments{
\item{id}{as required by shiny modules}
}
\description{
Allow user to set or change their user password.
Only shown if password can be set, i.e. user has been identified.
}
