% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_PIP_QualityImprovement.R
\name{qim}
\alias{qim}
\title{Quality Improvement user interface}
\usage{
qim(input, output, session, dMQIM, contact)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMQIM}{dMeasure QIM R6 object}

\item{contact}{(logical) TRUE if using 'contact' list
'contact' list uses active contact methods
FALSE if using 'appointment' list}
}
\value{
none
}
\description{
Quality Improvement user interface
}
