% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_PIP_QualityImprovement.R
\name{qim_UI}
\alias{qim_UI}
\title{qim_UI - Practice Incentive Program Quality Improvement Measures module}
\usage{
qim_UI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
PIP QI module
}
