% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables_definitions.R
\docType{data}
\name{semantic_popupJS}
\alias{semantic_popupJS}
\title{Javascript support code for semantic/fomantic JS popups}
\format{An object of class \code{character} of length 1.}
\usage{
semantic_popupJS
}
\description{
necessary for semantic/fomantic JS popups. included directly in datatables options
}
\details{
provide padding for export/print buttons
}
\keyword{datasets}
