% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_server.R
\name{servers_datatable}
\alias{servers_datatable}
\title{server configuration module - server}
\usage{
servers_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
list of database servers, name of chosen server
access to EMR database, access to configuration database}
}
\value{
count increments with each edit of server database
}
\description{
Server selector (dropdown)
Editable datatable with list of servers
}
