% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration.R
\name{servers_datatableUI}
\alias{servers_datatableUI}
\title{server configuration module - user interface function}
\usage{
servers_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Server selector (dropdown)
Editable datatable with list of servers
}
