% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{setPassword}
\alias{setPassword}
\title{setPassword}
\usage{
setPassword(newpassword, UserConfig, LoggedInUser, config_pool)
}
\arguments{
\item{newpassword}{the new password}

\item{UserConfig}{reactive, the User Configuration table}

\item{LoggedInUser}{reactive, the current user}

\item{config_pool}{reactive, access to the user configuration database}
}
\value{
nothing
}
\description{
sets password for user
Adjusts both the 'in-memory' data (UserConfig)
and the UserConfig database
}
