% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{simple_tag}
\alias{simple_tag}
\title{Simple tagger}
\usage{
simple_tag(msg, key = NULL)
}
\arguments{
\item{msg}{the text to decode}

\item{key}{the cipher, which can be set manually, otherwise will read from env}
}
\value{
- the hash
}
\description{
Simple tagger of text strings, will output a text string.
Uses sodium library and base64enc. Has some defaults, but
will also take command-line arguments or read from environment.
}
