% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{simple_tag_compare}
\alias{simple_tag_compare}
\title{Simple tag comparison}
\usage{
simple_tag_compare(msg, tag, key = NULL)
}
\arguments{
\item{msg}{the text to check}

\item{tag}{the tagged message (base64 encoded)}

\item{key}{the cipher, which can be set manually, otherwise will read from env}
}
\value{
- TRUE if matching, FALSE otherwise
}
\description{
Simple tagger of text strings, will output a text string.
Uses sodium library and base64enc. Has some defaults, but
will also take command-line arguments or read from environment.
}
