% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_appointment_df}
\alias{steps_appointment_df}
\title{appointment view help}
\usage{
steps_appointment_df()
}
\value{
returns a datatable of rintrojs steps
}
\description{
returns a dataframe of rintrojs steps
}
\details{
rintrojs steps for appointment view
}
