% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_cancerscreen_df}
\alias{steps_cancerscreen_df}
\title{cancer screening view help}
\usage{
steps_cancerscreen_df()
}
\value{
returns a datatable of rintrojs steps
}
\description{
returns a dataframe of rintrojs steps
}
\details{
rintrojs steps for cancer screen view
}
