% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_choose_clinician_date_df}
\alias{steps_choose_clinician_date_df}
\title{steps_choose_clinician_date_df}
\usage{
steps_choose_clinician_date_df()
}
\description{
returns a dataframe of rintrojs steps
}
\details{
rintrojs steps for choosing provider and date
}
