% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_choose_contact_details_df}
\alias{steps_choose_contact_details_df}
\title{steps_choose_contact_details_df}
\usage{
steps_choose_contact_details_df()
}
\value{
a dataframe of rintrojs steps
}
\description{
returns a dataframe of rintrojs steps
}
\details{
rintrojs steps for choosing contact specification
}
