% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_conditions_df}
\alias{steps_conditions_df}
\title{conditions help}
\usage{
steps_conditions_df(condition_tab = "Post-natal")
}
\arguments{
\item{condition_tab}{current condition tab which is open}
}
\value{
returns a datatable of rintrojs steps
}
\description{
returns a dataframe of rintrojs steps
}
\details{
rintrojs steps for conditions
}
