% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_immunization_df}
\alias{steps_immunization_df}
\title{immunizatioin view help}
\usage{
steps_immunization_df()
}
\value{
returns a datatable of rintrojs steps
}
\description{
returns a dataframe of rintrojs steps
}
\details{
rintrojs steps for immunization view
}
