% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_overview_df}
\alias{steps_overview_df}
\title{overview help}
\usage{
steps_overview_df()
}
\value{
returns a datatable of rintrojs steps
}
\description{
returns a dataframe of rintrojs steps
}
\details{
rintrojs steps for overview
}
