% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_users.R
\name{userconfig_datatable}
\alias{userconfig_datatable}
\title{user configuration module - server function}
\usage{
userconfig_datatable(input, output, session, UserConfig, UserFullConfig,
  UserRestrictions, LocationNames, db, config_pool)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{UserConfig}{reactiveval, list of user config}

\item{UserRestrictions}{reactiveval, user restriction list}

\item{LocationNames}{list of location names (not including ID or Description)}

\item{db}{reactivevalues link to EMR database. includes $users and $dbversion}

\item{config_pool}{reactiveval, access to configuration database}
}
\value{
count - increments with each GUI edit of user configuration database
}
\description{
user configuration module - server function
}
