% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_users.R
\name{userconfig_datatableUI}
\alias{userconfig_datatableUI}
\title{userconfig_datatableUI - editable datatable module}
\usage{
userconfig_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
user configuration module - user interface function
}
