% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_users.R
\name{userconfig_enableRestrictions}
\alias{userconfig_enableRestrictions}
\title{userconfig_enableRestrictions - server component}
\usage{
userconfig_enableRestrictions(input, out, session, UserConfig,
  UserRestrictions, config_pool)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{UserConfig}{reactiveval, list of user config}

\item{UserRestrictions}{reactiveval, list of user restrictions}

\item{config_pool}{reactiveval, access to configuration database}

\item{output}{as required by Shiny modules}
}
\value{
nothing
}
\description{
set or unset restrictions to user abilities
}
