% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_users.R
\name{userconfig_enableRestrictionsUI}
\alias{userconfig_enableRestrictionsUI}
\title{userconfig_enableRestrictionsUI}
\usage{
userconfig_enableRestrictionsUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
shiny user interface element
}
\description{
set or unset user restrictions
}
