% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_users.R
\name{userconfig_resetpasswordUI}
\alias{userconfig_resetpasswordUI}
\title{userconfig_resetpasswordUI}
\usage{
userconfig_resetpasswordUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
shiny user interface element
}
\description{
reset password of selected user
}
