% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Immunization.R
\name{vax_datatableUI}
\alias{vax_datatableUI}
\title{immunization module - UI function}
\usage{
vax_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Display immunization opportunities within selected range of dates and providers
}
