% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Appointments.R
\name{appointments_datatable}
\alias{appointments_datatable}
\title{appointment list module - server}
\usage{
appointments_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{- dMeasure R6 object}
}
\value{
none
}
\description{
list of appointments within selected range of dates and providers
}
