% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Billings.R
\name{billings_datatable}
\alias{billings_datatable}
\title{appointment list module - server}
\usage{
billings_datatable(input, output, session, dMBillings)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMBillings}{dMeasureBillings R6 object}
}
\value{
none
}
\description{
list of appointments and billings
within selected range of dates and providers
}
