% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CancerScreen.R
\name{cancerscreen_datatable}
\alias{cancerscreen_datatable}
\title{Cancer screening module server}
\usage{
cancerscreen_datatable(input, output, session, dM)
}
\arguments{
\item{input}{(as required by modules)}

\item{output}{(as required by modules)}

\item{session}{(as required by modules)}

\item{dM}{dMeasure R6 object}
}
\value{
None
}
\description{
Chronic disease management items claimed, pending or unclaimed for appointment list
}
