% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CDM.R
\name{cdm_datatable}
\alias{cdm_datatable}
\title{Chronic disease management list module - server}
\usage{
cdm_datatable(input, output, session, dMCDM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMCDM}{dMeasureCDM R6 object
access to appointments lists, condition lists, history and EMR database}
}
\value{
none
}
\description{
chronic disease management items claimed, pending or unclaimed for appointment list
}
