% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Conditions.R
\name{conditions_postnatal_datatable}
\alias{conditions_postnatal_datatable}
\title{condition post-natal - server}
\usage{
conditions_postnatal_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object
access to appointments lists, visits, user interface elements
and post-natal status database}
}
\value{
none
}
\description{
search for post-natal (or potentially post-natal) patients.
}
\details{
Filtered by number of days post-natal, whether a 'visit' recorded by the
selected physicians during the potential pregnancy period, and any
recorded pregnancy outcome.

Attached to appointments within the defined appointment search period.
}
