% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fomantic_definitions.R
\name{datatable_styled}
\alias{datatable_styled}
\title{styled DT datatable}
\usage{
datatable_styled(data, fillContainer = TRUE, extensions = c("Buttons",
  "Scroller", "Responsive"), dom = "frltiBp",
  buttons = list("copyHtml5", "print", list(extend = "collection",
  buttons = list(list(extend = "csvHtml5", filename = "DailyMeasure"),
  list(extend = "excel", filename = "DailyMeasure"), list(extend = "pdf",
  filename = "DailyMeasure")), text = "Download")),
  initComplete = DT::JS(semantic_popupJS), paging = FALSE,
  scrollY = "60vh", scrollX = FALSE, fixedColumns = FALSE, ...)
}
\arguments{
\item{initComplete}{by default, execute JS code to enable semantic popups}

\item{paging}{set to FALSE, otherwise hard to 'print'/copy a whole table}
}
\value{
DT datatable object
}
\description{
by default, have export/print buttons, only render what is visible.
javascript code to attach labels to semantic/fomantic JS popups.
no pagination.
}
