% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fomantic_definitions.R
\name{datatable_styled}
\alias{datatable_styled}
\title{styled DT datatable}
\usage{
datatable_styled(data, fillContainer = TRUE, extensions = c("Buttons",
  "Scroller", "Responsive"), dom = "frltiBp", buttons = list("colvis",
  list(extend = "copyHtml5", exportOptions = list(columns = ":visible")),
  list(extend = "print", exportOptions = list(columns = ":visible")),
  list(extend = "collection", buttons = list(list(extend = "csvHtml5",
  exportOptions = list(columns = ":visible"), filename = "DailyMeasure"),
  list(extend = "excel", exportOptions = list(columns = ":visible"),
  filename = "DailyMeasure"), list(extend = "pdf", exportOptions =
  list(columns = ":visible"), filename = "DailyMeasure")), text =
  "Download")), drawCallback = DT::JS(semantic_popupJS),
  `responsive-resize` = DT::JS(semantic_popupJS),
  `responsive-display` = DT::JS(paste("function ( e, datatable, row, showHide, update )",
  "{console.log( 'Details for row '+row.index()+' '",
  "+(showHide ? 'shown' : 'hidden'))}")), paging = FALSE,
  scrollY = "60vh", scrollX = FALSE, fixedColumns = FALSE,
  columnDefs = list(list()), ...)
}
\arguments{
\item{paging}{set to FALSE, otherwise hard to 'print'/copy a whole table}

\item{initComplete}{by default, execute JS code to enable semantic popups}
}
\value{
DT datatable object
}
\description{
by default, have export/print buttons, only render what is visible.
javascript code to attach labels to semantic/fomantic JS popups.
no pagination.
}
