% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatables_definitions.R
\name{datatable_styled}
\alias{datatable_styled}
\title{styled DT datatable}
\usage{
datatable_styled(
  data,
  fillContainer = TRUE,
  extensions = c("Buttons", "Scroller", "Responsive"),
  dom = "frltiBp",
  colvis = list(extend = "colvis", collectionLayout = "fixed four-column"),
  copyHtml5 = list(extend = "copyHtml5", exportOptions = list(columns = ":visible")),
  printButton = list(extend = "print", exportOptions = list(columns = ":visible")),
  downloadButton = list(extend = "collection", buttons = list(list(extend = "csvHtml5",
    exportOptions = list(columns = ":visible"), filename = "DailyMeasure"), list(extend =
    "excel", exportOptions = list(columns = ":visible"), filename = "DailyMeasure"),
    list(extend = "pdf", exportOptions = list(columns = ":visible"), filename =
    "DailyMeasure")), text = "Download", collectionLayout = "fixed"),
  drawCallback = DT::JS(semantic_popupJS),
  `responsive-resize` = DT::JS(semantic_popupJS),
 
    `responsive-display` = DT::JS(paste("function ( e, datatable, row, showHide, update )",
    "{console.log( 'Details for row '+row.index()+' '",
    "+(showHide ? 'shown' : 'hidden'))}")),
  paging = FALSE,
  scrollY = "60vh",
  deferRender = TRUE,
  scrollX = FALSE,
  fixedColumns = FALSE,
  columnDefs = list(list()),
  ...
)
}
\arguments{
\item{data}{the datatable}

\item{fillContainer}{TRUE to automatically fill containing element}

\item{colvis}{column visibility button. set to NULL if not wanted}

\item{copyHtml5}{copy button. set to NULL if not wanted}

\item{printButton}{print button. set to NULL if not wanted}

\item{downloadButton}{collection of download buttons. set to NULL if not wanted}

\item{drawCallback}{by default, execute JS code to enable semantic popups}

\item{paging}{'pages' for long table. default is FALSE, otherwise hard to 'print'/copy a whole table}

\item{scrollY}{percentage of window height}

\item{deferRender}{don't render all rows on page immediately}

\item{scrollX}{left-right scrolling}

\item{fixedColumns}{fix left- or right-most columns in place}

\item{columnDefs}{allow customization of columns e.g. hiding by default}
}
\value{
DT datatable object
}
\description{
by default, have export/print buttons, only render what is visible.
javascript code to attach labels to semantic/fomantic JS popups.
no pagination.
}
