% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fomantic_definitions.R
\name{datatable_styled}
\alias{datatable_styled}
\title{styled DT datatable}
\usage{
datatable_styled(data, fillContainer = TRUE, extensions = c("Buttons",
  "Scroller", "Responsive"), dom = "frltiBp", buttons = c("copyHtml5",
  "csvHtml5", "excel", "pdf", "print"),
  initComplete = DT::JS(semantic_popupJS), paging = FALSE,
  scrollY = "60vh", ...)
}
\description{
by default, have export/print buttons, only render what is visible.
javascript code to attach labels to semantic/fomantic JS popups.
no pagination.
}
