% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_CancerScreen.R
\name{fobt_list}
\alias{fobt_list}
\title{Bowel cancer screening list}
\usage{
fobt_list(appointments_list, emrpool)
}
\arguments{
\item{appointments_list}{reactive, list of appointments to search}

\item{emrpool}{accesss to Best Practice (EMR) database}
}
\value{
list of appointments (with patient details)
}
\description{
Bowel cancer screening list
}
