% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_location.R
\name{locations_datatable}
\alias{locations_datatable}
\title{location configuration module - server}
\usage{
locations_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object}
}
\value{
count - increments with each edit of server database
}
\description{
Editable datatable with list of locations
}
