% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_password.R
\name{passwordConfig_server}
\alias{passwordConfig_server}
\title{password configuration module - server function}
\usage{
passwordConfig_server(input, output, session, UserConfig, LoggedInUser,
  config_db)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{UserConfig}{reactiveval, list of user config}

\item{LoggedInUser}{reactiveval, currently logged in user}

\item{config_db}{R6 object, access to configuration database}
}
\value{
count - increments with each GUI edit of user configuration database
}
\description{
Allow user to set or change their user password
}
