% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_PIP_QualityImprovement.R
\name{qim}
\alias{qim}
\title{Quality Improvement user interface}
\usage{
qim(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 object}
}
\value{
none
}
\description{
Quality Improvement user interface
}
