% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_PIP_QualityImprovement.R
\name{qim_15plus}
\alias{qim_15plus}
\title{Quality Improvement fifteen plus (15+ years) - server}
\usage{
qim_15plus(input, output, session, dMQIM, contact)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMQIM}{dMeasure QIM R6 object
access to appointments lists, results, correspondence and EMR database}

\item{contact}{(logical) TRUE if using 'contact' list
'contact' list uses active contact methods
FALSE if using 'appointment' list}
}
\value{
none
}
\description{
Quality Improvement fifteen plus (15+ years) - server
}
