% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fomantic_definitions.R
\name{semantic_button}
\alias{semantic_button}
\title{Create semantic/fomantic buttons with attached tooltips (text and HTML)}
\usage{
semantic_button(button, colour = "", popuptext = NA, popuphtml = NA)
}
\arguments{
\item{button}{list of buttons contents}

\item{colour}{list of colours}

\item{popuptext}{list of popup texts}

\item{popuphtml}{(alternative) list of popup html}
}
\value{
vector of semantic/fomantic buttons
}
\description{
user-defined colour and popuptext (tooltip) or popuphtml (HTML tooltip)
}
