% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fomantic_definitions.R
\name{semantic_tag}
\alias{semantic_tag}
\title{Create semantic/fomantic tags with attached tooltips (text and HTML)}
\usage{
semantic_tag(tag, colour = "", popuptext = NA, popuphtml = NA)
}
\description{
Returns a vector of semantic/fomantic tags.
}
\details{
User-defined colour and popuptext (tooltip) or popuphtml (HTMl tooltip)
}
