% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_server.R
\name{servers_datatable}
\alias{servers_datatable}
\title{server configuration module - server}
\usage{
servers_datatable(input, output, session, BPdatabase, BPdatabaseChoice,
  emr_db, config_db)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{BPdatabase}{reactiveval, list of servers}

\item{BPdatabaseChoice}{reactiveval, name of chosen server}

\item{emr_db}{R6 object, current Best Practice (electronic medical record 'EMR') database in use}

\item{config_db}{R6 object, access to configuration database}
}
\value{
count increments with each edit of server database
}
\description{
Server selector (dropdown)
Editable datatable with list of servers
}
