% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_server.R
\name{servers_datatable}
\alias{servers_datatable}
\title{server configuration module - server}
\usage{
servers_datatable(input, output, session, BPdatabase, BPdatabaseChoice,
  emrpool, config_pool)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{BPdatabase}{reactiveval, list of servers}

\item{BPdatabaseChoice}{reactiveval, name of chosen server}

\item{emrpool}{reactiveval, current Best Practice (electronic medical record 'EMR') database pool in use}

\item{config_pool}{reactiveval, access to configuration database}
}
\value{
count increments with each edit of server database
}
\description{
Server selector (dropdown)
Editable datatable with list of servers
}
