% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduction.R
\name{steps_datatable_helpers}
\alias{steps_datatable_helpers}
\title{steps_datatable_helpers}
\usage{
steps_datatable_helpers(element_name, print_copy_view = TRUE)
}
\arguments{
\item{element_name}{name of the highlighted table element (with 'extra' print/copy view toggle)}

\item{print_copy_view}{include 'Print and Copy View' instructions (not all datatables have this)}
}
\value{
returns a datatable of rintrojs steps
}
\description{
returns a dataframe of rintrojs steps

steps for a typical datatable display which includes
1. column visibility
2. print/copy view
3. Copy button (only print/copy view)
4. Download/export button (only print/copy view)
5. search dialog
}
\details{
rintrojs steps for describing datatable view helpers
}
