% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration.R
\name{userconfig_datatableUI}
\alias{userconfig_datatableUI}
\title{user configuration module - user interface function}
\usage{
userconfig_datatableUI(id)
}
\arguments{
\item{id}{module ID (used in conjunction with 'callModule')}
}
\value{
Shiny user interface element
}
\description{
Editable datatable with list of servers
}
