% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Configuration_users.R
\name{userconfig_resetpassword}
\alias{userconfig_resetpassword}
\title{userconfig_resetpassword - server component}
\usage{
userconfig_resetpassword(input, output, session, UserConfig, config_db)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{UserConfig}{reactiveval, list of user config}

\item{config_db}{R6 object, access to configuration database}
}
\value{
nothing
}
\description{
reset password of selected user
}
