% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_Immunization.R
\name{vax_datatable}
\alias{vax_datatable}
\title{immunization module - UI function}
\usage{
vax_datatable(input, output, session, dM)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dM}{dMeasure R6 objection}
}
\value{
none
}
\description{
vaccinations done, pending or never done for appointment list
}
